move dynamic API from Term into a Mixin class
